web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "django-ninja>=1.6.2",
]
